Completed cards shuffling and winning/losing functionality
Implemented a shuffleCard function to shuffle cards each time one of the cards is clicked or when the difficulty level has been chosen. Also, whenever a card is clicked twice, the game ends and displays the Defeat OutcomeCard. If user won, it display the Victory OutcomeCard. When clicking the restart button, the game restarts by making all the wasClicked property to false, closing the OutcomeCard and opening the MainMenu. For some reason, I haven't implemented the functionality where all wasClicked property becomes false, but it works...