Addedhidden overflow in MainBody
Whenever the outcomeStatus state value is truthy (meaning when usre has lost or won and the outcomeCard is shown), the overflow proeprty of the body element will be set to hidden with useEffect hook. That is to disable user from scrolling the screen when the outcomeCard is shown. When the outcomeCard is hidden, the bodie's overflow property is set to auto.