Replaced svg files for png files for icon-images
Did that because I suspect that svg image formats are why the icon-images are not displaying. Now if you are wondering why I say suspect, I will assume that you are assuming that I didn't check first if they really are the culprit by using npm run preview or npm run build to check if they work in the dist folder, but both of these commands don't work for some reason.